requirements.txt
requests>=2.31.0
dnspython>=2.4.2
python-whois>=0.9.4
rich>=13.7.0
python-dotenv>=1.0.0
